+++ src/models/verification_oauth.rs
@@ -29,7 +29,7 @@ pub struct VerificationOauth {
 	)]
 	pub error: Option<Option<Box<models::ClerkError>>>,
 	#[serde(rename = "expire_at")]
-	pub expire_at: i32,
+	pub expire_at: i64,
 	#[serde(rename = "attempts", deserialize_with = "Option::deserialize")]
 	pub attempts: Option<i32>,
 	#[serde(
@@ -42,7 +42,7 @@ pub struct VerificationOauth {
 }
 
 impl VerificationOauth {
-	pub fn new(status: Status, strategy: String, expire_at: i32, attempts: Option<i32>) -> VerificationOauth {
+	pub fn new(status: Status, strategy: String, expire_at: i64, attempts: Option<i32>) -> VerificationOauth {
 		VerificationOauth {
 			object: None,
 			status,
