+++ src/models/verification_passkey.rs
@@ -31,7 +31,7 @@ pub struct VerificationPasskey {
 	#[serde(rename = "attempts", deserialize_with = "Option::deserialize")]
 	pub attempts: Option<i32>,
 	#[serde(rename = "expire_at", deserialize_with = "Option::deserialize")]
-	pub expire_at: Option<i32>,
+	pub expire_at: Option<i64>,
 	#[serde(
 		rename = "verified_at_client",
 		default,
@@ -42,7 +42,7 @@ pub struct VerificationPasskey {
 }
 
 impl VerificationPasskey {
-	pub fn new(status: Status, strategy: Strategy, attempts: Option<i32>, expire_at: Option<i32>) -> VerificationPasskey {
+	pub fn new(status: Status, strategy: Strategy, attempts: Option<i32>, expire_at: Option<i64>) -> VerificationPasskey {
 		VerificationPasskey {
 			object: None,
 			status,
